Prepare for search.
I want to add search functionality to Go Read. My first approach is to
add a new API endpoint that accepts a searchterm and returns all stories
that include that searchterm.

For larger amounts of feeds this probably needs to be done with a
kind of paging, so I don't return hundreds of feeds at once, but for
now I'm more concerned about getting to a place where I can run Queries
on the datastore at all...